test(core): split candidate spatial and store characterization suites
Replace the 539-line spatial and 501-line store monoliths with focused
suites under tests/candidate/spatial/ and tests/candidate/store/,
sharing largePointScene fixtures for the spatial hot-path guards (#504).
